Выделимл компонент film-promo отдельно от main
