+++ src/hocs/with-main/with-main.jsx
@@ -41,30 +41,6 @@ const withMain = (Component) => {
       PropTypes.arrayOf(PropTypes.node),
       PropTypes.node,
     ]),
-    handleOnPlayClick: PropTypes.func.isRequired,
-    filmPromo: PropTypes.shape({
-      id: PropTypes.number,
-      title: PropTypes.string,
-      poster: PropTypes.string,
-      genre: PropTypes.string,
-      dateRelease: PropTypes.number,
-      cover: PropTypes.string,
-      videoSrc: PropTypes.string,
-      previewVideoSrc: PropTypes.string,
-      director: PropTypes.string,
-      actors: PropTypes.arrayOf(PropTypes.string),
-      ratingScore: PropTypes.number,
-      ratingCount: PropTypes.number,
-      description: PropTypes.string,
-      runTime: PropTypes.number,
-      backgroundImage: PropTypes.string,
-      reviews: PropTypes.arrayOf(PropTypes.shape({
-        comment: PropTypes.string,
-        rating: PropTypes.number,
-        author: PropTypes.string,
-        date: PropTypes.string,
-      })),
-    }),
     films: PropTypes.arrayOf(PropTypes.shape({
       id: PropTypes.number.isRequired,
       title: PropTypes.string.isRequired,
